requirements.txt
python-dotenv
beautifulsoup4
pandas
openpyxl
lxml
html5lib
numpy
matplotlib
